requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
inference-sdk==0.9.15
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1